requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.6
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.24.0.post1
